feat: support html game
